requirements.txt
datasus-dbc==0.1.3
dbfread==2.0.7
dotenv-python==0.0.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1
shell==1.0.1
